Documentation: Change `first-steps` examples structure
