+++ docs/src/tutorial/first-steps-final/Cargo.toml
@@ -1,7 +1,7 @@
 # Copyright Kani Contributors
 # SPDX-License-Identifier: Apache-2.0 OR MIT
 [package]
-name = "kani-first-steps"
+name = "first-steps-final"
 version = "0.1.0"
 edition = "2018"
